factor out ContactsList from invite friends
